Perform sandsurfacerenderer update after loading calibration
We need to force an update of the conversion matrices.
Normally this happens inside the update loop of ofApp, but the
KinectProjector update routine is run first, which clears the
calibrationUpdated flag. If this flag isn't set, the conversion
matrices will be updated, hence we need to force an update before this
flag can be cleared.